Moves ParticipantConfig and dependants to `participant-integration-api`
changelog_begin
changelog_end
