+++ ledger/ledger-runner-common/src/main/scala/com/daml/ledger/runner/common/CliParticipantConfig.scala
@@ -4,6 +4,7 @@
 package com.daml.ledger.runner.common
 
 import com.daml.lf.data.Ref
+import com.daml.platform.config.ParticipantRunMode
 import com.daml.platform.configuration.IndexServiceConfig
 import com.daml.platform.indexer.IndexerConfig
 import com.daml.ports.Port
