+++ ledger/sandbox-on-x/src/main/scala/com/daml/ledger/sandbox/SandboxOnXRunner.scala
@@ -48,6 +48,7 @@ import scala.concurrent.{ExecutionContext, ExecutionContextExecutorService}
 import scala.util.chaining._
 import com.daml.ledger.configuration.LedgerId
 import com.daml.ledger.runner.common.MetricsConfig.MetricRegistryType
+import com.daml.platform.config.{ParticipantConfig, ParticipantRunMode}
 import com.daml.platform.store.DbSupport.ParticipantDataSourceConfig
 import com.daml.ports.Port
 import com.daml.platform.store.interning.StringInterningView
